fix random color function

+++ Sources/HsinUtils/UIColorRandomColorExtension.swift
@@ -9,6 +9,6 @@ import UIKit
 
 public extension UIColor {
   static func makeRandomColor() -> UIColor {
-    return UIColor(displayP3Red: CGFloat(Int.random(in: 0...255)/255), green: CGFloat(Int.random(in: 0...255)/255), blue: CGFloat(Int.random(in: 0...255)/255), alpha: 1.0)
+    return UIColor(displayP3Red: CGFloat(Int.random(in: 0...255))/CGFloat(255), green: CGFloat(Int.random(in: 0...255))/CGFloat(255), blue: CGFloat(Int.random(in: 0...255))/CGFloat(255), alpha: 1.0)
   }
 }
